requirements.txt
uvicorn==0.35.0
fastapi==0.115.14
supabase==2.16.0
python-dotenv==1.1.1
httpx==0.28.1
